requirements.txt
absl-py==2.1.0
click==8.1.7
colorama==0.4.6
joblib==1.4.2
lxml==5.3.0
nltk==3.6.5
numpy==1.23.5
pandas==2.1.0
portalocker==2.10.1
python-dateutil==2.9.0.post0
python-docx==0.8.11
pytz==2024.1
regex==2024.7.24
rouge==1.0.1
rouge_score==0.1.2
sacrebleu==2.4.3
six==1.16.0
tabulate==0.9.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
